requirements.txt

fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.25.0
cachetools>=5.3.0
structlog>=23.2.0
email-validator>=2.1.0
python-dotenv>=1.0.0

# ========================================
# 🧪 Development & Testing (Optional)
# ========================================
# Comment out these for production to reduce build time
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# pytest-cov>=4.1.0
# pytest-mock>=3.12.0

# ========================================
# 🔧 Development Tools (Optional)
# ========================================
# Comment out these for production
# black>=23.0.0
# flake8>=6.0.0
# isort>=5.12.0
# mypy>=1.0.0
